requirements.txt
alembic==1.6.5
beautifulsoup4==4.9.3
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
colorama==0.4.4
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
Flask-Login==0.5.0
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
heroku==0.1.4
idna==3.2
itsdangerous==2.0.1
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==2.0.1
psycopg2==2.9.1
python-dateutil==1.5
python-decouple==3.4
python-dotenv==0.19.0
python-editor==1.0.4
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.3.23
urllib3==1.26.6
Werkzeug==2.0.1
wincertstore==0.2
WTForms==2.3.3
